fix: clean up project structure and delete unnecessary files
