--- Form1.cs
@@ -1,40 +1,0 @@
-namespace MARK200
-{
-    public partial class Form1 : Form
-    {
-        public Form1()
-        {
-            InitializeComponent();
-        }
-
-        private void panel2_Paint(object sender, PaintEventArgs e)
-        {
-
-        }
-
-        private void textBox1_TextChanged(object sender, EventArgs e)
-        {
-
-        }
-
-        private void panel1_Paint(object sender, PaintEventArgs e)
-        {
-
-        }
-
-        private void label2_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void checkBoxShowPassword_CheckedChanged(object sender, EventArgs e)
-        {
-            textBoxPassword.UseSystemPasswordChar = !checkBoxShowPassword.Checked;
-        }
-
-        private void textBoxPassword_TextChanged(object sender, EventArgs e)
-        {
-            textBoxPassword.UseSystemPasswordChar = true;
-        }
-    }
-}
