+++ TriQue/Forms/LoginForm.cs
@@ -1,10 +1,40 @@
-namespace TriQue
+namespace TriQue.Forms
 {
     public partial class LoginForm : Form
     {
         public LoginForm()
         {
             InitializeComponent();
+        }
+
+        private void panel2_Paint(object sender, PaintEventArgs e)
+        {
+
+        }
+
+        private void textBox1_TextChanged(object sender, EventArgs e)
+        {
+
+        }
+
+        private void panel1_Paint(object sender, PaintEventArgs e)
+        {
+
+        }
+
+        private void label2_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void checkBoxShowPassword_CheckedChanged(object sender, EventArgs e)
+        {
+            textBoxPassword.UseSystemPasswordChar = !checkBoxShowPassword.Checked;
+        }
+
+        private void textBoxPassword_TextChanged(object sender, EventArgs e)
+        {
+            textBoxPassword.UseSystemPasswordChar = true;
         }
     }
 }
