+++ TriQue/Program.cs
@@ -1,4 +1,5 @@
 using TriQue.Data;
+using TriQue.Forms;
 
 
 namespace TriQue
