--- TriQue_Login.slnx
@@ -1,3 +1,0 @@
-<Solution>
-  <Project Path="TriQue_Login.csproj" />
-</Solution>
